+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 - Added Change Log
 - Removed `ScriptableEnum<T>` class
 - Unified `GetByName()`logic
+- Added the same `_PREDEFINED_COUNT` to `countThreshold` in two methods